+++ openeo/capabilities.py
@@ -4,7 +4,7 @@ import re
 from typing import Union, Tuple
 
 
-# Is this base class (still) useful?
+# TODO Is this base class (still) useful?
 
 
 class Capabilities(ABC):
@@ -117,6 +117,9 @@ class ComparableVersion:
 
     def __str__(self):
         return ".".join(map(str, self._version))
+
+    def __hash__(self):
+        return hash(self._version)
 
     def to_string(self):
         return str(self)