+++ tests/rest/datacube/test_datacube.py
@@ -4,9 +4,9 @@ General cube method tests against both
 - 1.0.0-style DataCube
 
 """
-
-from datetime import date, datetime
 import pathlib
+from datetime import date, datetime
+from unittest import mock
 
 import numpy as np
 import pytest
@@ -16,9 +16,10 @@ import shapely.geometry
 from openeo.capabilities import ComparableVersion
 from openeo.rest import BandMathException
 from openeo.rest.datacube import DataCube
-from .conftest import API_URL
-from .. import get_download_graph
+
 from ... import load_json_resource
+from .. import get_download_graph
+from .conftest import API_URL
 
 
 def test_apply_dimension_temporal_cumsum(s2cube, api_version):
@@ -446,3 +447,192 @@ def test_download_bytes(connection, requests_mock, api_version, format, expected
     requests_mock.post(API_URL + '/result', content=result_callback)
     result = connection.load_collection("S2").download(format=format)
     assert result == b"data"
+
+
+class TestExecuteBatch:
+    @pytest.fixture
+    def get_create_job_pg(self, connection):
+        """Fixture to help intercepting the process graph that was passed to Connection.create_job"""
+        with mock.patch.object(connection, "create_job") as create_job:
+
+            def get() -> dict:
+                assert create_job.call_count == 1
+                return create_job.call_args[1]["process_graph"]
+
+            yield get
+
+    def test_create_job_defaults(self, s2cube, get_create_job_pg, recwarn, caplog):
+        s2cube.create_job()
+        pg = get_create_job_pg()
+        assert set(pg.keys()) == {"loadcollection1", "saveresult1"}
+        assert pg["saveresult1"] == {
+            "process_id": "save_result",
+            "arguments": {
+                "data": {"from_node": "loadcollection1"},
+                "format": "GTiff",
+                "options": {},
+            },
+            "result": True,
+        }
+        assert recwarn.list == []
+        assert caplog.records == []
+
+    @pytest.mark.parametrize(
+        ["out_format", "expected"],
+        [("GTiff", "GTiff"), ("NetCDF", "NetCDF")],
+    )
+    def test_create_job_out_format(
+        self, s2cube, get_create_job_pg, out_format, expected
+    ):
+        s2cube.create_job(out_format=out_format)
+        pg = get_create_job_pg()
+        assert set(pg.keys()) == {"loadcollection1", "saveresult1"}
+        assert pg["saveresult1"] == {
+            "process_id": "save_result",
+            "arguments": {
+                "data": {"from_node": "loadcollection1"},
+                "format": expected,
+                "options": {},
+            },
+            "result": True,
+        }
+
+    @pytest.mark.parametrize(
+        ["save_result_format", "execute_format", "expected"],
+        [
+            ("GTiff", "GTiff", "GTiff"),
+            ("GTiff", None, "GTiff"),
+            ("NetCDF", "NetCDF", "NetCDF"),
+            ("NetCDF", None, "NetCDF"),
+        ],
+    )
+    def test_create_job_existing_save_result(
+        self,
+        s2cube,
+        get_create_job_pg,
+        save_result_format,
+        execute_format,
+        expected,
+    ):
+        cube = s2cube.save_result(format=save_result_format)
+        cube.create_job(out_format=execute_format)
+        pg = get_create_job_pg()
+        assert set(pg.keys()) == {"loadcollection1", "saveresult1"}
+        assert pg["saveresult1"] == {
+            "process_id": "save_result",
+            "arguments": {
+                "data": {"from_node": "loadcollection1"},
+                "format": expected,
+                "options": {},
+            },
+            "result": True,
+        }
+
+    @pytest.mark.parametrize(
+        ["save_result_format", "execute_format"],
+        [("NetCDF", "GTiff"), ("GTiff", "NetCDF")],
+    )
+    def test_create_job_existing_save_result_incompatible(
+        self, s2cube, save_result_format, execute_format
+    ):
+        cube = s2cube.save_result(format=save_result_format)
+        with pytest.raises(ValueError):
+            cube.create_job(out_format=execute_format)
+
+    def test_execute_batch_defaults(self, s2cube, get_create_job_pg, recwarn, caplog):
+        s2cube.execute_batch()
+        pg = get_create_job_pg()
+        assert set(pg.keys()) == {"loadcollection1", "saveresult1"}
+        assert pg["saveresult1"] == {
+            "process_id": "save_result",
+            "arguments": {
+                "data": {"from_node": "loadcollection1"},
+                "format": "GTiff",
+                "options": {},
+            },
+            "result": True,
+        }
+        assert recwarn.list == []
+        assert caplog.records == []
+
+    @pytest.mark.parametrize(
+        ["out_format", "expected"],
+        [("GTiff", "GTiff"), ("NetCDF", "NetCDF")],
+    )
+    def test_execute_batch_out_format(
+        self, s2cube, get_create_job_pg, out_format, expected
+    ):
+        s2cube.execute_batch(out_format=out_format)
+        pg = get_create_job_pg()
+        assert set(pg.keys()) == {"loadcollection1", "saveresult1"}
+        assert pg["saveresult1"] == {
+            "process_id": "save_result",
+            "arguments": {
+                "data": {"from_node": "loadcollection1"},
+                "format": expected,
+                "options": {},
+            },
+            "result": True,
+        }
+
+    @pytest.mark.parametrize(
+        ["output_file", "expected"],
+        [("cube.tiff", "GTiff"), ("cube.nc", "netCDF")],
+    )
+    def test_execute_batch_out_format_from_output_file(
+        self, s2cube, get_create_job_pg, output_file, expected
+    ):
+        s2cube.execute_batch(outputfile=output_file)
+        pg = get_create_job_pg()
+        assert set(pg.keys()) == {"loadcollection1", "saveresult1"}
+        assert pg["saveresult1"] == {
+            "process_id": "save_result",
+            "arguments": {
+                "data": {"from_node": "loadcollection1"},
+                "format": expected,
+                "options": {},
+            },
+            "result": True,
+        }
+
+    @pytest.mark.parametrize(
+        ["save_result_format", "execute_format", "expected"],
+        [
+            ("GTiff", "GTiff", "GTiff"),
+            ("GTiff", None, "GTiff"),
+            ("NetCDF", "NetCDF", "NetCDF"),
+            ("NetCDF", None, "NetCDF"),
+        ],
+    )
+    def test_execute_batch_existing_save_result(
+        self,
+        s2cube,
+        get_create_job_pg,
+        save_result_format,
+        execute_format,
+        expected,
+    ):
+        cube = s2cube.save_result(format=save_result_format)
+        cube.execute_batch(out_format=execute_format)
+        pg = get_create_job_pg()
+        assert set(pg.keys()) == {"loadcollection1", "saveresult1"}
+        assert pg["saveresult1"] == {
+            "process_id": "save_result",
+            "arguments": {
+                "data": {"from_node": "loadcollection1"},
+                "format": expected,
+                "options": {},
+            },
+            "result": True,
+        }
+
+    @pytest.mark.parametrize(
+        ["save_result_format", "execute_format"],
+        [("NetCDF", "GTiff"), ("GTiff", "NetCDF")],
+    )
+    def test_execute_batch_existing_save_result_incompatible(
+        self, s2cube, save_result_format, execute_format
+    ):
+        cube = s2cube.save_result(format=save_result_format)
+        with pytest.raises(ValueError):
+            cube.execute_batch(out_format=execute_format)